+++ CHANGES.md
@@ -1,9 +1,10 @@
 ## Change Log
 
 #### 0.2.0
-* Adds codenarc for Groovy projects
 * Adds JDepend for Java and Groovy projects
 * Adds FindBugs for Java and Groovy projects
+* Improve license header checks
+* Adds CodeNarc for Groovy projects
 
 #### 0.1.1
 * Adds prependJavaHeader task for each source set